make undo snackbar duration short

+++ app/src/main/java/com/godzuche/dend/features/main/impl/presentation/MainScreen.kt
@@ -179,7 +179,7 @@ fun MainScreen(
                     actionLabel = actionLabel,
                     duration = if (actionLabel == null) {
                         SnackbarDuration.Short
-                    } else SnackbarDuration.Long,
+                    } else SnackbarDuration.Short,
                 )
 
                 when (result) {
